+++ modules/Base.js
@@ -1,5 +1,6 @@
 /* eslint-disable no-underscore-dangle */
 const fs = require('fs').promises;
+const { join } = require('path');
 const winston = require('winston');
 
 class Base {
@@ -12,12 +13,19 @@ class Base {
   }
 
   /**
+   * In case of logging sometimes we might need to replace name
+   */
+  getConstructorName() {
+    return this.constructor.name;
+  }
+
+  /**
    * Optimzation to lazy load logger. It will be inited only on request
    */
   get logger() {
     if (!this._realLogger) {
       this._realLogger = this.getLogger(
-        this.constructor.loggerGroup + this.constructor.name,
+        this.constructor.loggerGroup + this.getConstructorName(),
       );
     }
     return this._realLogger;
@@ -72,21 +80,48 @@ class Base {
   }
 
   async loadFilesWithInheritance(internalFolder, externalFolder) {
+    this.logger.warn(
+      'Method "loadFilesWithInheritance" deprecated. Please use "getFilesPathWithInheritance"',
+    );
+    return (
+      await this.getFilesPathWithInheritance(internalFolder, externalFolder)
+    ).map((file) => file.path);
+  }
+
+  async getFilesPathWithInheritance(internalFolder, externalFolder) {
+    async function rreaddir(dir, allFiles = []) {
+      const files = (await fs.readdir(dir)).map((f) => join(dir, f));
+      allFiles.push(...files);
+      await Promise.all(
+        files.map(async (f) => {
+          if ((await fs.stat(f)).isDirectory()) {
+            allFiles.pop();
+            return rreaddir(f, allFiles);
+          }
+          return null;
+        }),
+      );
+      return allFiles.map((file) => file.replace(`${dir}/`, ''));
+    }
+
     let [internalFiles, externalFiles] = await Promise.all([
-      fs.readdir(internalFolder),
-      fs.readdir(externalFolder),
+      rreaddir(internalFolder),
+      rreaddir(externalFolder),
     ]);
 
     const filterIndexFile = (fileName) => {
+      const fileArray = fileName.split('/');
+      const file = fileArray[fileArray.length - 1];
       return (
-        fileName[0] === fileName[0].toUpperCase() &&
-        fileName[0] !== '.' &&
-        !fileName.includes('.test.js')
+        file[0] === file[0].toUpperCase() && // Start with capital
+        file[0] !== '.' && // not start with dot
+        !file.includes('.test.js') // not test files
       );
     };
 
     internalFiles = internalFiles.filter(filterIndexFile);
     externalFiles = externalFiles.filter(filterIndexFile);
+
     const filesToLoad = [];
     for (const file of internalFiles) {
       if (externalFiles.includes(file)) {
@@ -94,12 +129,18 @@ class Base {
           `Skipping register INTERNAL file ${file} as it override by EXTERNAL ONE`,
         );
       } else {
-        filesToLoad.push(`${internalFolder}/${file}`);
+        filesToLoad.push({
+          path: `${internalFolder}/${file}`,
+          file,
+        });
       }
     }
 
     for (const file of externalFiles) {
-      filesToLoad.push(`${externalFolder}/${file}`);
+      filesToLoad.push({
+        path: `${externalFolder}/${file}`,
+        file,
+      });
     }
     return filesToLoad;
   }